# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Documentation",
    "version": "1.0.0",
    "description": "API documentation for your Express.js app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://localhost:50000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "fetchPost": {
        "type": "object",
        "properties": {
          "post": {
            "$ref": "#/components/schemas/getPost"
          },
          "qset": {
            "$ref": "#/components/schemas/getQsetArray"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/fetchPost"
            }
          }
        }
      },
      "getPost": {
        "type": "object",
        "properties": {
          "id_post": {
            "$ref": "#/components/schemas/id"
          },
          "title": {
            "$ref": "#/components/schemas/title"
          },
          "content": {
            "$ref": "#/components/schemas/content"
          },
          "publish_date": {
            "$ref": "#/components/schemas/publish_date"
          },
          "publisher": {
            "$ref": "#/components/schemas/username"
          },
          "plays": {
            "$ref": "#/components/schemas/amount"
          },
          "signals": {
            "$ref": "#/components/schemas/amount"
          }
        }
      },
      "newPost": {
        "type": "object",
        "properties": {
          "title": {
            "$ref": "#/components/schemas/title"
          },
          "content": {
            "$ref": "#/components/schemas/content"
          },
          "qsetArray": {
            "$ref": "#/components/schemas/postQsetArray"
          }
        }
      },
...
      "newReply": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/content"
          }
        }
      },
      "getQsetArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/qsetNoAns"
        }
      },
      "postQsetArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/qsetNoID"
        }
      },
...
      "qsetNoAns": {
        "type": "object",
        "properties": {
          "id_set": {
            "$ref": "#/components/schemas/id"
          },
          "prompt": {
            "$ref": "#/components/schemas/prompt"
          },
          "ressource_type": {
            "$ref": "#/components/schemas/ressource_type"
          },
          "ressource_link": {
            "$ref": "#/components/schemas/link"
          },
          "questionArray": {
            "$ref": "#/components/schemas/questionArrayNoAns"
          }
        }
      },
      "qsetNoID": {
        "type": "object",
        "properties": {
          "prompt": {
            "$ref": "#/components/schemas/prompt"
          },
          "ressource_type": {
            "$ref": "#/components/schemas/ressource_type"
          },
          "ressource_link": {
            "$ref": "#/components/schemas/link"
          },
          "questionArray": {
            "$ref": "#/components/schemas/questionArrayNoID"
          }
        }
      },
      "answerRepSubArr": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/answerRepSub"
        }
      },
      "answerRepSub": {
        "type": "object",
        "properties": {
          "id_set": {
            "$ref": "#/components/schemas/id"
          },
          "id_question": {
            "$ref": "#/components/schemas/id"
          },
          "right": {
            "$ref": "#/components/schemas/is_correct"
          }
        }
      },
      "answerSubArr": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/answerSub"
        }
      },
      "answerSub": {
        "type": "object",
        "properties": {
          "id_set": {
            "$ref": "#/components/schemas/id"
          },
          "id_question": {
            "$ref": "#/components/schemas/id"
          },
          "is_correct": {
            "$ref": "#/components/schemas/is_correct"
          }
        }
      },
      "questionArrayNoAns": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/questionNoAns"
        }
      },
      "questionArrayNoID": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/questionNoID"
        }
      },
      "questionNoAns": {
        "type": "object",
        "properties": {
          "id_set": {
            "$ref": "#/components/schemas/id"
          },
          "id_question": {
            "$ref": "#/components/schemas/id"
          },
          "statement": {
            "$ref": "#/components/schemas/statement"
          }
        }
      },
      "questionNoID": {
        "type": "object",
        "properties": {
          "is_correct": {
            "$ref": "#/components/schemas/is_correct"
          },
          "statement": {
            "$ref": "#/components/schemas/statement"
          }
        }
      },
      "q": {
        "type": "string",
        "example": "Query%20Example",
        "description": "The query is url encoded"
      },
      "gs": {
        "type": "string",
        "example": "0%201%208",
        "description": "The list is separated by space %20"
      },
      "username": {
        "type": "string",
        "example": "myusername"
      },
...
      "id": {
        "type": "integer",
        "example": 1
      },
      "title": {
        "type": "string",
        "example": "Hello World !"
      },
      "content": {
        "type": "string",
        "example": "My name is Nico."
      },
      "is_correct": {
        "type": "boolean",
        "example": true
      },
      "statement": {
        "type": "string",
        "example": "A example of an affirmation."
      },
      "prompt": {
        "type": "string",
        "example": "What is the correct answer ?"
      },
      "ressource_type": {
        "type": "string",
        "example": "image"
      },
      "link": {
        "type": "string",
        "example": "https://cdn.example.com/image.jpg"
      },
      "publish_date": {
        "type": "string",
        "format": "date-time",
        "example": "2017-07-21T17:32:28Z"
      },
      "amount": {
        "type": "integer",
        "example": 169
      },
...
      "signaled": {
        "type": "boolean",
        "example": true
      },
...
    },
    "responses": {
      "401": {
        "description": "Unauthorized, must provide an authentification bearer jwt"
      },
      "403": {
        "description": "Although your token is valid, you don't the rights to do so"
      }
    }
  },
  "paths": {
...
    "/post": {
      "get": {
        "description": "Query some posts",
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "gs",
            "description": "The gamescope is selection of the game you want to search for. The list is separated by dash -",
            "schema": {
              "$ref": "#/components/schemas/gs"
            }
          },
          {
            "in": "query",
            "name": "q",
            "description": "The query (q) parameter is based text you want to search for in the title.",
            "schema": {
              "$ref": "#/components/schemas/q"
            }
          },
          {
            "in": "query",
            "name": "u",
            "description": "The user parameter is to select post for only one username.",
            "schema": {
              "$ref": "#/components/schemas/username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the result from db",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/getPost"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad information provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "description": "Post a new post",
        "tags": [
          "Post"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/newPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New post created !"
          },
          "400": {
            "description": "Bad information provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/post/{postId}": {
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "required": true,
          "description": "The post Id to manage"
        }
      ],
      "get": {
        "description": "Fetch a specific post, its replies and its questionss",
        "tags": [
          "Post"
        ],
        "responses": {
          "200": {
            "description": "Return the result from db",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/fetchPost"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "put": {
        "tags": [
          "Post"
        ],
        "description": "Modify a a specific post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/newReply"
                  },
                  {
                    "$ref": "#/components/schemas/newPost"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reply modified !"
          },
          "400": {
            "description": "Bad information provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "delete": {
        "description": "Delete a specific post and its replies",
        "tags": [
          "Post"
        ],
        "responses": {
          "200": {
            "description": "Reply deleted !"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/post/{postId}/reply": {
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "required": true,
          "description": "The post Id to manage"
        }
      ],
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Reply to a specific post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/newReply"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New reply created !"
          },
          "400": {
            "description": "Bad information provided"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/post/{postId}/submit": {
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "required": true,
          "description": "The post Id to manage"
        }
      ],
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Reply to a specific post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/answerSubArr"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission sucessfull ! Please find attached your results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/answerRepSubArr"
                }
              }
            }
          },
          "400": {
            "description": "Bad length"
          },
          "404": {
            "description": "Post not found"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/post/{postId}/signal": {
      "parameters": [
        {
          "in": "path",
          "name": "postId",
          "schema": {
            "$ref": "#/components/schemas/id"
          },
          "required": true,
          "description": "The post Id to manage"
        }
      ],
      "get": {
        "description": "Get amount of signals of a specific post",
        "tags": [
          "Post"
        ],
        "responses": {
          "200": {
            "description": "Return the result from db",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "$ref": "#/components/schemas/amount"
                    },
                    "signaled": {
                      "$ref": "#/components/schemas/signaled"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "post": {
        "description": "User signal a specific post",
        "tags": [
          "Post"
        ],
        "responses": {
          "201": {
            "description": "New signal created !"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}